+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVP8MKSSRF2F10WA8WZW8A5G"
+        "01JVPKVJD9ABHGMK3CTSTHXXYR"
     )
     r"""The id of the resource"""