+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVP8MKSQJPM2S116KQGHEFWZ"
+        "01JVPKVJD6GNBQB09FMSXGQ79P"
     )
     r"""The id of the resource"""