+++ packages/orq-rc/RELEASES.md
@@ -1309,3 +1309,13 @@ Based on:
 - [python v3.6.0-rc.16] packages/orq-rc
 ### Releases
 - [PyPI v3.6.0-rc.16] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.16 - packages/orq-rc
+
+## 2025-05-20 10:30:56
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.548.1 (2.605.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.6.0-rc.17] packages/orq-rc
+### Releases
+- [PyPI v3.6.0-rc.17] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.17 - packages/orq-rc